+++ services/seller_frontend_service/select_ad_reactor_web_test.cc
@@ -215,11 +215,11 @@ TYPED_TEST(SelectAdReactorForWebTest, VerifyChaffedResponse) {
   EXPECT_TRUE(deserialized_auction_result->is_chaff());
 }
 
-auto EqLogContext(const LogContext& log_context) {
-  return AllOf(
-      Property(&LogContext::generation_id, Eq(log_context.generation_id())),
-      Property(&LogContext::adtech_debug_id,
-               Eq(log_context.adtech_debug_id())));
+auto EqLogContext(const server_common::LogContext& log_context) {
+  return AllOf(Property(&server_common::LogContext::generation_id,
+                        Eq(log_context.generation_id())),
+               Property(&server_common::LogContext::adtech_debug_id,
+                        Eq(log_context.adtech_debug_id())));
 }
 
 auto EqGetBidsRawRequestWithLogContext(